Changed led_num type to uint32_t

+++ include/structs.h
@@ -26,7 +26,7 @@ typedef struct LEDDict
     CRGBSet led_array;
     uint8_t *hue_array;
     uint8_t *led_args;
-    uint16_t led_num;
+    uint32_t led_num;
     uint32_t current_mode;
     uint32_t previous_mode;
     uint32_t current_time;
