+++ src/effects/rainbow_wave.h
@@ -24,7 +24,7 @@
 class RainbowWave
 {
 public:
-    static void run(CRGBSet &led_array, uint8_t hue_array[], uint16_t led_num, bool &effect_setup, uint32_t refresh_delay, uint32_t &current_time);
+    static void run(CRGBSet &led_array, uint8_t hue_array[], uint32_t led_num, bool &effect_setup, uint32_t refresh_delay, uint32_t &current_time);
 };
 
 #endif // RAINBOWWAVE_H
